feat: User registration API
Added the model and DTO required for registration.
Passwords are encrypted using bcrypt.

If email or username are taken, a custom exceptions is thrown. Reasoning: we could rely on built in integrity exceptions, but they are domain-bound.

I fixed an issue with the Session object binding to SQLAlchemy instead of SQLModel (SQLAlchemy does not have an exec() function).