feat: Initial login API
This is the initial implementation. There are additional requirements that need to be considered, some of which are discussed in Issue #4.

The JWTBearer class is triggered for each HTTP request to handle authorization.

Currently, the JWT token contains the following attributes: id, role, must_change_password, and expires. More attributes will be added soon.

The login and test controllers have been verified to confirm the functionality of the current setup.

Using a DTO for a response with a single attribute might be excessive and needs to be reviewed.