feat: Implement `get repos of user` API
Get repos created by user + repos in orgs the user is a member of.
If repo is private, hide it, unless the user sending the API request has the proper credentials.

Access control is done in the service layer because the logic will get even more complex once we implement Teams.