feat: Add client-side search+filter for `get repos of user`
It's client-side because 1) it's easier and 2) we can assume that one user will never have so many repositories that we'd need to paginate and implement filtering server-side. I say <100 repos per user is _fine_. No need to future-proof.

Just in case, we can first cache and only later change the architecture and API.