test: Add dummy integration test for ElasticSearch
I've added the `_z_` in the filename to assure that this test is run last. The reason we _want_ to run these tests last is because ElasticSearch is most likely to have started by the time we reach these tests, so the chance of timeout (i.e. test fail) are as low as possible.