fix: Resolve `No such image` error caused by GitHub runner
I have tested on the mirrored repository and got an error, even though I committed only a comment and made a pull request.

The problem was caused by GitHub runners. The last test related to the previous PR #43 was on June 19th. The runner image used by GitHub Actions was ubuntu-24.04 (20250602.3.0) with Docker Compose v2.36.0.

Today, GitHub Actions used the runner image ubuntu-24.04 (20250622.1.0) with Docker Compose v2.36.2.

If I understand correctly, the following PR introduced the issue, as services listed in `depends_on` are no longer automatically built before running the backend service. Because of that, I first build all services, as it is the best way to ensure that we don’t encounter a problem when the backend is up with any of its `depends_on` services.

https://github.com/docker/compose/pull/12863

Locally, I didn’t have any issues with Docker Compose and running the exact same command, as the Docker Compose version on my PC is older and `build` services before `up`.

Used Ubuntu images:
https://github.com/actions/runner-images/blob/releases/ubuntu24/20250602/images/ubuntu/Ubuntu2404-Readme.md
https://github.com/actions/runner-images/blob/releases/ubuntu24/20250622/images/ubuntu/Ubuntu2404-Readme.md